video-player/requirements.txt
#numpy==1.15.4
#opencv-python==3.4.3.18
#Pillow==5.3.0
numpy
opencv-python
Pillow